+++ util/src/main/scala/org/broadinstitute/dsde/workbench/util/package.scala
@@ -1,15 +1,23 @@
 package org.broadinstitute.dsde.workbench
 
 import java.io.FileInputStream
+import java.nio.file.Path
 
-import cats.effect.{Resource, Sync}
+import cats.effect.{Concurrent, ContextShift, Resource, Sync}
+import io.circe.Decoder
+import io.circe.fs2.decoder
+import fs2.{Stream, io}
+import org.typelevel.jawn.AsyncParser
+
+import scala.concurrent.ExecutionContext
+import scala.concurrent.ExecutionContext.global
 import scala.concurrent.duration._
 
 /**
  * Created by dvoet on 2/24/17.
  */
 package object util {
-  def toScalaDuration(javaDuration: java.time.Duration) = Duration.fromNanos(javaDuration.toNanos)
+  def toScalaDuration(javaDuration: java.time.Duration): FiniteDuration = Duration.fromNanos(javaDuration.toNanos)
 
   def addJitter(baseTime: FiniteDuration, maxJitterToAdd: Duration): FiniteDuration = {
     baseTime + ((scala.util.Random.nextFloat * maxJitterToAdd.toNanos) nanoseconds)
@@ -32,4 +40,17 @@ package object util {
 
   def readFile[F[_]](path: String)(implicit sf: Sync[F]): Resource[F, FileInputStream] =
     Resource.make(sf.delay(new FileInputStream(path)))(f => sf.delay(f.close()))
+
+  /*
+   * Example:
+   * scala> org.broadinstitute.dsde.workbench.util.readJsonFileToA[IO, List[String]](java.nio.file.Paths.get("/tmp/list"), None).compile.lastOrError.unsafeRunSync
+   * res0: List[String] = List(this is great)
+   *
+   */
+  def readJsonFileToA[F[_]: Sync: ContextShift: Concurrent, A: Decoder](path: Path, blockingExecutionContext: Option[ExecutionContext] = None): Stream[F, A] = {
+    io.file.readAll[F](path, blockingExecutionContext.getOrElse(global), 4096)
+      .through(fs2.text.utf8Decode)
+      .through(_root_.io.circe.fs2.stringParser(AsyncParser.SingleValue))
+      .through(decoder)
+  }
 }